fix: move synchronous utility out of Server Actions file to resolve build error
Moved \`normalizeMessageContent\` to \`lib/utils/index.ts\` as Next.js 15 requires all exports from 'use server' files to be async functions.
